removing download from root command and adding download command (main command)
