+++ cmd/root.go
@@ -22,30 +22,12 @@ package cmd
 
 import (
 	"fmt"
-	"net/http"
 	"os"
-	"strings"
-	"time"
 
 	"github.com/go-playground/log"
 	"github.com/go-playground/log/handlers/console"
 	"github.com/spf13/cobra"
 )
-
-//DownloadResult struct for structuring download results once
-//files are successfully downloaded or failed.
-type DownloadResult struct {
-	FileName, FileSize, Result, Error string
-	Attempt                           int
-	Duration                          time.Duration
-}
-
-var cfgFile string
-var debugStatus bool
-var salesForceUserName string
-var salesForceUserPassword string
-var maxWorkers int
-var httpClient *http.Client
 
 // rootCmd -> download org files (main purpose)
 var rootCmd = &cobra.Command{
@@ -53,37 +35,9 @@ var rootCmd = &cobra.Command{
 	Short: "Commandline application for downloading SalesForce backup files.",
 	Long: `A commandline application for downloading multiple files concurrently.
 You should execute giving Username (u) and Password (p). For example:
-go-salesforce-backup-downloader.exe -u sadmin@atyourcrazyorg -p mypasswordwithtoken
-go-salesforce-backup-downloader.exe --user sadmin@atyourcrazyorg --password mypasswordwithtoken
+go-salesforce-backup-downloader.exe download -u sadmin@atyourcrazyorg -p mypasswordwithtoken
+go-salesforce-backup-downloader.exe download --user sadmin@atyourcrazyorg --password mypasswordwithtoken
 `,
-
-	Run: func(cmd *cobra.Command, args []string) {
-		start := time.Now()
-		fmt.Printf("Start time -> %s\n", start.Format(time.ANSIC))
-		loginData := login()
-
-		var consolidateResults []DownloadResult
-
-		backupfolder := strings.Split(salesForceUserName, "@")[len(strings.Split(salesForceUserName, "@"))-1]
-
-		if _, err := os.Stat(backupfolder); os.IsNotExist(err) {
-			log.Infof("backupfolder folder doesn't exist. Creating: %s", backupfolder)
-			err := os.Mkdir(backupfolder, 0777)
-			if err != nil {
-				log.Fatal(err)
-			}
-		}
-
-		export(loginData, &consolidateResults)
-
-		//Export results
-		exportResultsToCsv(consolidateResults)
-
-		t := time.Now()
-		fmt.Printf("End time -> %s\n", t.Format(time.ANSIC))
-		fmt.Printf("total time -> %s\n", t.Sub(start))
-
-	},
 }
 
 // Execute adds all child commands to the root command and sets flags appropriately.
@@ -96,52 +50,7 @@ func Execute() {
 }
 
 func init() {
-	// Future development -> accept config file
-	//cobra.OnInitialize(initConfig)
 
 	cLog := console.New(true)
 	log.AddHandler(cLog, log.AllLevels...)
-
-	httpClient = &http.Client{
-		Timeout: time.Minute * 10,
-	}
-
-	// global parameters for the application.
-	// Future development -> accept config file
-	// rootCmd.PersistentFlags().StringVar(&cfgFile, "config", "", "config file (default is $HOME/.go-salesforce-backup-downloader.yaml)")
-	rootCmd.PersistentFlags().StringVarP(&salesForceUserName, "user", "u", "", "Salesforce username e.g something@somewhere.there")
-	rootCmd.PersistentFlags().StringVarP(&salesForceUserPassword, "password", "p", "", "Salesforce password+token e.g supersecretpasswordwithtoken")
-	rootCmd.PersistentFlags().IntVarP(&maxWorkers, "maxworkers", "m", 5, "Maximum number of workers for concurrency. (default is 5)")
-	rootCmd.MarkPersistentFlagRequired("user")
-	rootCmd.MarkPersistentFlagRequired("password")
-
-	// debug flag should be use for verbosity
-	//rootCmd.PersistentFlags().BoolVarP(&debugStatus, "debug", "d", false, "debug?")
 }
-
-// initConfig reads in config file and ENV variables if set.
-// currently no config file is being passed.
-// func initConfig() {
-// 	if cfgFile != "" {
-// 		// Use config file from the flag.
-// 		viper.SetConfigFile(cfgFile)
-// 	} else {
-// 		// Find home directory.
-// 		home, err := homedir.Dir()
-// 		if err != nil {
-// 			fmt.Println(err)
-// 			os.Exit(1)
-// 		}
-
-// 		// Search config in home directory with name ".go-salesforce-backup-downloader" (without extension).
-// 		viper.AddConfigPath(home)
-// 		viper.SetConfigName(".go-salesforce-backup-downloader")
-// 	}
-
-// 	viper.AutomaticEnv() // read in environment variables that match
-
-// 	// If a config file is found, read it in.
-// 	if err := viper.ReadInConfig(); err == nil {
-// 		fmt.Println("Using config file:", viper.ConfigFileUsed())
-// 	}
-//}
